Add site parse script, story retrieve script. Update memento and timemap script to scale by months.

+++ mementos.py
@@ -2,9 +2,12 @@ import requests
 import hashlib
 import json
 import os
-# import csv
-# from datetime import datetime
-# from urllib.parse import urlparse
+from calendar import monthrange
+
+
+months = ["201605", "201606", "201607", "201608", "201609", "201610",
+          "201611",
+          "201612", "201701", "201702", "201703", "201704", "201705"]
 
 
 def get_mementos(day, uri):
@@ -28,30 +31,39 @@ def makedir_if_not_exists(directory):
         os.makedirs(directory)
 
 
-def get_news_mementos():
+def get_news_mementos(mo):
     """
     Get html for websites listed in textfile.
     """
     with open('data/news-websites.txt') as f:
 
+        out_mo = mo[:4] + "_" + mo[4:]
         for news_site in f:
             saved_mementos = {}
             news_site = news_site.rstrip()
             site_hash = hashlib.md5(news_site.encode()).hexdigest()
             saved_mementos.setdefault(site_hash, {})
-            makedir_if_not_exists("data/mementos/" + site_hash)
-            # memento_pairs = []
-            # November 1 - 30
-            if os.path.isfile("data/mementos/{}/mementos.json".
-                              format(site_hash)):
-                print("{} already has mementos. Delete folder to download "
-                      "mementos again".format(site_hash))
+            makedir_if_not_exists(
+                "data/mementos/{}/{}".format(site_hash, out_mo))
+            # URI hash, formatted folder name
+            summary_file = "data/mementos/{}/{}/mementos.json".format(
+                site_hash, out_mo)
+
+            if os.path.isfile(summary_file):
+                print("{} already has mementos for {}. Delete folder to"
+                      " download mementos again".format(news_site, out_mo))
                 continue
 
-            for day in list(range(1, 31)):
+            # Iterate through days in the month
+            num_days = monthrange(int(mo[:4]), int(mo[4:]))[1]
+            for day in list(range(1, num_days + 1)):
                 day = "%02d" % (day,)
-                uri = "http://web.archive.org/web/201611{}010000if_/{}".format(
-                    day, news_site)
+                outfile = "data/mementos/{}/{}/{}.html".format(
+                    site_hash, out_mo, day)
+                if os.path.isfile(outfile):
+                    continue
+                uri = "http://web.archive.org/web/{}{}010000if_/{}".format(
+                    mo, day, news_site)
                 resp = get_mementos(day, uri)
 
                 # map request URI, response URI, and memento datetime
@@ -60,14 +72,23 @@ def get_news_mementos():
                           "Memento-Datetime": resp.headers["Memento-Datetime"]
                           })
                 print("Response:", resp.url)
-                with open("data/mementos/{}/{}.html".
-                          format(site_hash, day), 'w') as out:
+                with open(outfile, 'w') as out:
                     out.write(resp.text)
-
-            with open("data/mementos/{}/mementos.json".format(site_hash), 'w')\
-                    as memento_index:
-                json.dump(saved_mementos, memento_index)
+            #
+            # WRITE TO UPDATE ALREADY WRITTEN SUMMARY FILE WITH SAV
+            #
+            if os.path.isfile(summary_file):
+                data = {}
+                with open(summary_file, 'r') as memento_index:
+                    data = json.load(memento_index)
+                data[site_hash].update(saved_mementos[site_hash])
+                with open(summary_file, 'w') as memento_index:
+                    json.dump(data, memento_index)
+            else:
+                with open(summary_file, 'w') as memento_index:
+                    json.dump(saved_mementos, memento_index)
 
 
 if __name__ == "__main__":
-    get_news_mementos()
+    for mo in months:
+        get_news_mementos(mo)
